+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     - SAPBY for the BlockDataContainer now does not require an `out` to be passed (#2008)
     - Fixed the rendering of the SAG/SAGA documentation (#2011)
     - Set aliases: ISTA=PGD, FISTA=APGD (#2007)
+    - Add accelerated version to TransmissionAbsorption processor (#2032)
   - Dependencies:
     - Added scikit-image to CIL-Demos conda install command as needed for new Callbacks notebook.
     - Replaced matplotlib dependency with matplotlib-base (#2031)